requirements.txt
behave
python-dotenv
boto3
fordev
Pillow
tqdm
aioboto3
pandas
scikit-image
plotly
tensorflow
Flask
schedule
pymongo